# source/openapi.yaml
openapi: 3.1.0
info:
  title: localLLM-FastAPI
  description: |
    FastAPI経由で、localLLMを動かすスケーラブルなAPIアプリケーション

    ## 機能
    - 🔄 テキスト生成とエコー
    - ❤️ ヘルスチェック
    - 🌐 外部API統合（モック）
    - 📖 自動生成ドキュメント
    - 🔧 TypeScript型生成対応

    ## 開発者向け情報
    本APIは手書きのOpenAPI YAMLスキーマから自動生成されたコードベースです。
    スキーマファーストの開発アプローチを採用しています。
  version: 1.0.0
  contact:
    name: localLLM-FastAPI開発チーム
    url: https://github.com/ForLearnOrganization/localllm-fastapi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 開発サーバー
  - url: http://127.0.0.1:8000
    description: ローカル開発サーバー
...
paths:
...
  /api/v1/text/generate:
    post:
      tags: [text]
      summary: テキスト生成
      description: ルールベースまたはLLMを使用したテキスト生成
      operationId: generate_text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTextRequest"
            example:
              prompt: "こんにちは世界"
              max_length: 100
              temperature: 0.7
      responses:
        "200":
          description: テキスト生成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTextResponse"
              example:
                generated_text: "こんにちは世界！今日は素晴らしい日ですね。"
                input_prompt: "こんにちは世界"
                metadata:
                  method: "rule_based"
                  length: 25
        "422":
          description: 入力検証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /generate:
    post:
      tags: [text]
      summary: テキスト生成（後方互換）
      description: 既存コードとの後方互換性のためのエンドポイント
      operationId: generate_text_legacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTextRequest"
      responses:
        "200":
          description: テキスト生成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateTextResponse"
components:
  schemas:
...
    GenerateTextRequest:
      type: object
      properties:
        prompt:
          type: string
          description: テキスト生成用のプロンプト
          example: "今日の天気は"
        max_length:
          type: integer
          description: 生成テキストの最大長
          default: 100
          minimum: 1
          maximum: 1000
        temperature:
          type: number
          description: テキスト生成の温度パラメータ
          default: 0.7
          minimum: 0.0
          maximum: 2.0
      required:
        - prompt
    GenerateTextResponse:
      type: object
      properties:
        generated_text:
          type: string
          description: 生成されたテキスト
        input_prompt:
          type: string
          description: 元の入力プロンプト
        metadata:
          type: object
          properties:
            method:
              type: string
              description: 生成手法
              enum: [rule_based, llm]
            length:
              type: integer
              description: 生成テキストの文字数
            generation_time:
              type: number
              description: 生成にかかった時間（秒）
      required:
        - generated_text
        - input_prompt
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: エラーの詳細
        error_code:
          type: string
          description: エラーコード
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻
      required:
        - detail